Move MachineEncoder into machine.py
